+++ inst/mod/r/core.r
@@ -1,0 +1,8 @@
+#' @export
+box::use(
+    methods[...],
+    stats[...],
+    graphics[...],
+    grDevices[...],
+    utils[...]
+)