+++ tests/testthat/mod/nested/rel_import.r
@@ -1,15 +1,17 @@
+box::use(utils[tail])
+
 #' @export
-global_path = box:::mod_search_path(environment())
+global_path = tail(box:::mod_search_path(environment()), 1L)
 
 #' @export
 path_in_fun = function () {
-    box:::mod_search_path(environment())
+    tail(box:::mod_search_path(environment()), 1L)
 }
 
 #' @export
 path_in_nested_fun = function () {
     f = function () {
-        box:::mod_search_path(environment())
+        tail(box:::mod_search_path(environment()), 1L)
     }
     f()
 }